Transform the Keyword Builder in an external js
